auth-service/requirements.txt
asn1crypto==1.4.0
bcrypt==3.2.0
boto3==1.17.104
botocore==1.20.104
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
ecdsa==0.17.0
Flask==2.0.1
Flask-Cors==3.0.10
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.18.0
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.6
uuid==1.30
Werkzeug==2.0.1
